Update name of installable package
