+++ setup.cfg
@@ -1,20 +1,22 @@
 [metadata]
-name = nomenclature
+name = openentrance
 author = openENTRANCE consortium
 author_email = huppmann@iiasa.ac.at
-license = Apache License 2.0
-description = Model linkage nomenclature for the openENTRANCE project
+license = APACHE-2.0
+description = Nomenclature for the openENTRANCE project: variables, units, regions, ...
 long_description = file: README.md
 long_description_content_type = text/x-md
 url = https://github.com/openENTRANCE/nomenclature
 
 [options]
-packages = nomenclature
+packages = openentrance
 include_package_data = True
 install_requires =
     setuptools >= 41
     pyyaml
-    pyam-iamc  # the pyam package is released on pypi under this name
+    nomenclature
+    pyam-iamc >= 1.0  # the pyam package is released on pypi under this name
+    iam-units >= 2021.11.12
 setup_requires =
     setuptools >= 41
     setuptools_scm
